Added more support for streamlining tool changes.
It is now possible to define chunks of G-code that should be executed
when the machine pauses on an M6 and after it is unpaused again. This
can be used to automate moving to a tool change position.

The tool change dialog also now displays which tool is being switched
to (the Txx on the M6 line) so one doesn't have to bring up the serial
console to see this.

Finally, the "reposition" button in the tool change dialog now includes
resetting the coordinate system (g54 etc) that was in force when the M6
was hit, in case the user changes coordinate systems as part of the tool
change.